requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
fonttools==4.58.2
frozenlist==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
jiter==0.10.0
kiwisolver==1.4.8
lazy_loader==0.4
magic-filter==1.0.12
matplotlib==3.10.3
multidict==6.4.4
networkx==3.4.2
numpy==2.2.6
openai==1.84.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.0
aiomysql
